+++ src/main/java/com/example/starsign/database/DatabaseSource.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 class DatabaseSource(
-    override val cardid:Int,
+    override val cardid:Long,
     override val title : String,
     val source: Map<Mana, Int>
 ): DatabaseCard(cardid, title) {}
